Refactor project structure: reorganize imports
