+++ main.py
@@ -1,4 +1,10 @@
-from src.Interface import launch_interface
+from src.DomainMind.Interface import launch_interface
+
+# Enable LangSmith tracing
+# Load environment variables from .env file to configure LangSmith 
+# API keys and settings
+from dotenv import load_dotenv
+load_dotenv()
 
 
 def main():
