+++ src/DomainMind/OutputProcess.py
